Refactor test
